Fix and refactor UmaSelectorDialog and CreateUmaDialog
